Don't create symlink if exists /etc/resolv.conf
FIX: if /etc/resolv.conf created other dhcp client then rc.firstboot finish with exit code and don't removed itself from rc.local.

+++ files/firstboot/25-create-resolv-symlink.sh
@@ -11,5 +11,7 @@ if [ ! -f "/run/systemd/resolve/resolv.conf" ] ; then
   touch /run/systemd/resolve/resolv.conf
 fi
 
-# Create symlink to /etc/reolv.conf
-ln -s /run/systemd/resolve/resolv.conf /etc/resolv.conf
+# Create symlink to /etc/reolv.conf if not exists yet
+if [ ! -e "/etc/resolv.conf" ] ; then
+  ln -s /run/systemd/resolve/resolv.conf /etc/resolv.conf
+fi
